Create InsertionSort.cpp
Insertion sorting using both array (uncomment the code) or linked lists.
This algorithm has a minimum O(n) and max O(n^2) in case of array and it is the same for lists.
It is an adaptive algorithm, and we do not need a flag like in bubble sort. 
Best case scenario the list is sorted, worst case scenario is in reverse order.
If the list is sorted then we do not swap anything so O(n)
It will also not shift the elements if they are equal, so it is also a stable algorithm as it leaves the duplicates in order after sorting.